pyproject.toml
[project]
name = "anki-inject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=0.3.0",
    "pillow>=12.0.0",
    "pyperclip>=1.11.0",
    "pygobject>=3.42.0",
    "python-dotenv>=1.2.1",
    "pywebview>=6.1",
    "httpx>=0.27.0",
    "psutil>=6.0.0",
    "memory-profiler>=0.61.0",
]
